+++ core/services/vrf/vrf_hash_to_curve_cost_test.go
@@ -44,7 +44,7 @@ func deployVRFContract(t *testing.T) (contract, common.Address) {
 	}
 	auth, _ := bind.NewKeyedTransactorWithChainID(&key, testutils.SimulatedChainID)
 	genesisData := core.GenesisAlloc{auth.From: {Balance: assets.Ether(100)}}
-	gasLimit := ethconfig.Defaults.Miner.GasCeil
+	gasLimit := uint32(ethconfig.Defaults.Miner.GasCeil)
 	backend := cltest.NewSimulatedBackend(t, genesisData, gasLimit)
 	parsed, err := abi.JSON(strings.NewReader(
 		solidity_vrf_verifier_wrapper.VRFTestHelperABI))
